unchi

+++ variables.h
@@ -14,7 +14,7 @@ int baseTime;       // フレームレート制御用変数
 int count;          // フレームレート制御用変数
 float fps;          // フレームレート保持変数
 char fpsString[256];
-int tmpAttack;      // 攻撃力計算用変数
+float tmpAttack;      // 攻撃力計算用変数
 bool win;           // 買ったか
 bool lose;          // 負けたか
 bool trueClear;     // 裏ボスを倒したか
